@@ -1,18 +1,19 @@
 ﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using BoltOn.Data.EF;
 using BoltOn.Samples.Application.Handlers;
-using BoltOn.Data.CosmosDb;
 using BoltOn.Bus.MassTransit;
-using BoltOn.Utilities;
 using BoltOn.Samples.Infrastructure.Data;
+using Microsoft.EntityFrameworkCore;
+using MassTransit;
+using System;
+using Microsoft.Extensions.Hosting;
 
 namespace BoltOn.Samples.WebApi
 {
-	public class Startup
+    public class Startup
 	{
 		public Startup(IConfiguration configuration)
 		{
@@ -23,22 +24,53 @@ namespace BoltOn.Samples.WebApi
 
 		public void ConfigureServices(IServiceCollection services)
 		{
-			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 			services.BoltOn(options =>
 			{
 				options.BoltOnEFModule();
-				options.BoltOnCosmosDbModule();
 				options.BoltOnMassTransitBusModule();
 				options.BoltOnCqrsModule();
-				options.BoltOnAssemblies(typeof(PingHandler).Assembly, typeof(SchoolDbContext).Assembly);
+				options.BoltOnAssemblies(typeof(PingHandler).Assembly, typeof(SchoolWriteDbContext).Assembly);
 			});
-		}
+			
+			var writeDbConnectionString = Configuration.GetValue<string>("SqlWriteDbConnectionString");
+            var readDbConnectionString = Configuration.GetValue<string>("SqlReadDbConnectionString");
+            var rabbitmqUri = Configuration.GetValue<string>("RabbitMqUri");
+			var rabbitmqUsername = Configuration.GetValue<string>("RabbitMqUsername");
+			var rabbitmqPassword = Configuration.GetValue<string>("RabbitMqPassword");
+            services.AddMassTransit(x =>
+            {
+                x.AddBus(provider => MassTransit.Bus.Factory.CreateUsingRabbitMq(cfg =>
+                {
+                    var host = cfg.Host(new Uri(rabbitmqUri), hostConfigurator =>
+                    {
+                        hostConfigurator.Username(rabbitmqUsername);
+                        hostConfigurator.Password(rabbitmqPassword);
+                    });
+                }));
+            });
 
-		public void Configure(IApplicationBuilder app, IApplicationLifetime appLifetime)
+            services.AddDbContext<SchoolWriteDbContext>(options =>
+            {
+                options.UseSqlServer(writeDbConnectionString);
+            });
+
+            services.AddDbContext<SchoolReadDbContext>(options =>
+            {
+                options.UseSqlServer(readDbConnectionString);
+            });
+
+        }
+
+		public void Configure(IApplicationBuilder app, IHostApplicationLifetime appLifetime)
 		{
-			app.UseMvc();
-			app.ApplicationServices.TightenBolts();
-			appLifetime.ApplicationStopping.Register(() => BoltOnAppCleaner.Clean());
+			app.UseRouting(); 
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
+            app.ApplicationServices.TightenBolts();
+			appLifetime.ApplicationStopping.Register(() => app.ApplicationServices.LoosenBolts());
 		}
 	}
 }
