@@ -5,8 +5,6 @@ using BoltOn.Cqrs;
 using BoltOn.Data;
 using BoltOn.Mediator.Pipeline;
 using BoltOn.Samples.Application.Entities;
-using Microsoft.Azure.Documents;
-using Microsoft.Azure.Documents.Client;
 
 namespace BoltOn.Samples.Application.Handlers
 {
@@ -34,11 +32,10 @@ namespace BoltOn.Samples.Application.Handlers
 
 		public async Task HandleAsync(StudentCreatedEvent request, CancellationToken cancellationToken)
 		{
-			var requestOptions = new RequestOptions { PartitionKey = new PartitionKey(request.StudentTypeId) };
-			var student = await _repository.GetByIdAsync(request.StudentId, requestOptions, cancellationToken);
+			var student = await _repository.GetByIdAsync(request.StudentId, cancellationToken: cancellationToken);
 
 			if (student == null)
-				await _repository.AddAsync(new StudentFlattened(request), cancellationToken);
+				await _repository.AddAsync(new StudentFlattened(request), cancellationToken: cancellationToken);
 		}
 	}
 }
