@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Reflection;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace BoltOn.Bootstrapping
 {
@@ -9,7 +10,16 @@ namespace BoltOn.Bootstrapping
 
         internal bool IsCqrsEnabled { get; set; }
 
-        public void BoltOnAssemblies(params Assembly[] assemblies)
+		internal List<object> OtherOptions { get; } = new List<object>();
+
+		public IServiceCollection ServiceCollection { get; private set; }
+
+		public BoltOnOptions(IServiceCollection serviceCollection)
+		{
+			ServiceCollection = serviceCollection;
+		}
+
+		public void BoltOnAssemblies(params Assembly[] assemblies)
 		{
 			AssembliesToBeIncluded.AddRange(assemblies);
 		}
